Website: convert `editorState` to js module
