+++ website/playground/composables/editor-state.js
@@ -1,0 +1,38 @@
+import { reactive, watch } from "vue";
+import * as storage from "../storage.js";
+
+const stateToggler = (property) => () => {
+  state[property] = !state[property];
+};
+
+const state = reactive({
+  showSidebar: window.innerWidth > window.innerHeight,
+  showAst: false,
+  showPreprocessedAst: false,
+  showDoc: false,
+  showComments: false,
+  showSecondFormat: false,
+  showInput: true,
+  showOutput: true,
+  rethrowEmbedErrors: false,
+  toggleSidebar: stateToggler("showSidebar"),
+  toggleAst: stateToggler("showAst"),
+  togglePreprocessedAst: stateToggler("showPreprocessedAst"),
+  toggleDoc: stateToggler("showDoc"),
+  toggleComments: stateToggler("showComments"),
+  toggleSecondFormat: stateToggler("showSecondFormat"),
+  toggleInput: stateToggler("showInput"),
+  toggleOutput: stateToggler("showOutput"),
+  toggleEmbedErrors: stateToggler("rethrowEmbedErrors"),
+  ...storage.get("editor_state"),
+});
+
+watch(
+  () => state,
+  () => {
+    storage.set("editor_state", state);
+  },
+  { deep: true },
+);
+
+export { state as editorState };
